make power scenario test portable
